ura-properties/requirements.txt
pandas==1.3.4
numpy==1.22.2
urllib3==1.26.7
boto3==1.21.12
awswrangler==2.14.0
onemapsg==1.0.3
PyYAML==6.0
